:recycle: optimize detail page
