+++ src/components/BestCombinationDetail/Header.tsx
@@ -1,3 +1,4 @@
+import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import Modal from '@components/Common/UI/Modal';
 import Like from '@components/Common/Button/Like';
@@ -5,7 +6,6 @@ import useLikedBestCombination from '@hooks/useLikedBestCombination';
 import styled from '@emotion/styled';
 import mediaQuery from '@styles/media-queries';
 import { changeRem, flexbox } from '@styles/mixin';
-import { useEffect } from 'react';
 
 type Tprops = {
   author: string;
